Implementando esquema de cálculo de risco para respiration monitor, notificação dos listeners, alterando o nível de listeners para cada variável. Alterando a estratégia de cálculo de risco.

+++ src/observer/HRObserver.java
@@ -2,6 +2,6 @@ package observer;
 
 public interface HRObserver {
 	
-	public void changeHR(float value);
+	public void changeHR(int value);
 
 }
